+++ src/Symfony/Component/DependencyInjection/ChildDefinition.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\DependencyInjection;
 
-use Symfony\Component\DependencyInjection\Exception\BadMethodCallException;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Exception\OutOfBoundsException;
 
@@ -104,21 +103,5 @@ class ChildDefinition extends Definition
         }
 
         return $this;
-    }
-
-    /**
-     * @internal
-     */
-    public function setAutoconfigured(bool $autoconfigured): self
-    {
-        throw new BadMethodCallException('A ChildDefinition cannot be autoconfigured.');
-    }
-
-    /**
-     * @internal
-     */
-    public function setInstanceofConditionals(array $instanceof): self
-    {
-        throw new BadMethodCallException('A ChildDefinition cannot have instanceof conditionals set on it.');
     }
 }